requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==4.2.5
filelock==3.12.3
platformdirs==3.10.0
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.24.5
django-encrypted-fields==2.3.0
psycopg2-binary==2.9.6 
pycryptodome==3.17
